Lokale html file gemaakt waar ik de backend functionaliteit in kan gaan testen
Deze commit is vooral zodat ik verder kan werken op een andere pc
