+++ project/index.html
@@ -1,0 +1,56 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>font-end test</title>
+</head>
+<body>
+    <h1>This is a test!</h1>
+    <button id="fetchBtn">Click me!</button>
+    <br><br>
+    <form id="myForm">
+        <input type="text" name="username" placeholder="Username">
+        <input type="password" name="password" placeholder="Password">
+        <button type="submit">Submit</button>
+    </form>
+    <script>
+        document.addEventListener("DOMContentLoaded", () => { // zorgt er voor dat de javascript pas wordt uitgevoerd nadat de html pagina volledig geladen is
+            const fetchBtn = document.getElementById('fetchBtn');
+            const form = document.getElementById('myForm');
+
+            form.addEventListener('submit', (e) => {
+                e.preventDefault();
+                console.log("testlog form");
+
+                const formData = new FormData(form);
+                console.log([...formData.entries()]); // deze notatie zorgt ervoor dat je de formData als een array kan loggen
+
+                const formDataObject = {};
+
+                // Zet de data van de form in een array
+                formData.forEach((value, key) => {
+                    formDataObject[key] = value;
+                });
+
+                // Zet de data van de array in een variabele
+                const username = formDataObject.username;
+                const password = formDataObject.password;
+
+                // Logt de variabelen
+                console.log("Username:", username);
+                console.log("Password:", password);
+            })
+
+            fetchBtn.addEventListener('click', () => {
+                console.log("button press registered")
+                fetch('/api/subscriptions')
+                    .then(res => res.json())
+                    .then(data => console.log(data));
+            });
+        })
+        
+                
+    </script>
+</body>
+</html>
